refactor: 아티클 카드 및 features 도메인 이름 note로 변경 [BETA-115]
- ArticleCard -> NoteCard
- features/notebook -> features/note
