+++ src/entities/note/ui/note-card/Meta.tsx
@@ -1,5 +1,5 @@
-import { ArticleMeta } from "@/shared/types/post.types";
-import * as S from "./ArticleCard.css";
+import { ArticleMeta } from "@/entities/note/note.types";
+import * as S from "./NoteCard.css";
 import HeartIcon from "@/shared/assets/icons/common/heart.svg";
 
 export default function Meta({ author, date, likeCount }: ArticleMeta) {
