+++ src/entities/note/ui/note-card/NoteCard.css.ts
